requirements.txt
google-generativeai>=0.8.3
openai-whisper>=20231117
ffmpeg-python>=0.2.0
torch>=2.0.0
numpy>=1.24.0
typing-extensions>=4.9.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pillow>=10.0.0
opencv-python>=4.12.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
boto3>=1.35.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
PyJWT>=2.9.0
yt-dlp>=2024.8.6
requests>=2.32.3
